+++ core/workflows/skybox.cpp
@@ -25,7 +25,7 @@ void SkyboxGraphics::createRenderPass()
         utils::Attachments::imageDescription(frame.bloom.format())
     };
 
-    utils::vkDefault::SubpassInfos subpassInfos = utils::vkDefault::subpassInfos(attachments.size());
+    utils::vkDefault::SubpassInfos subpassInfos = utils::vkDefault::subpassInfos(static_cast<uint32_t>(attachments.size()));
 
     utils::vkDefault::RenderPass::SubpassDependencies dependencies;
     dependencies.push_back(VkSubpassDependency{});
